Break message classes out into files
This file was getting out of control.
