+++ tchannel/mapping.py
@@ -1,24 +1,6 @@
 from __future__ import absolute_import
 
-from .messages import CallRequestMessage
-from .messages import CallResponseMessage
-from .messages import ErrorMessage
-from .messages import InitRequestMessage
-from .messages import InitResponseMessage
-from .messages import PingRequestMessage
-from .messages import PingResponseMessage
-
-
-ALL_MESSAGES = [
-    InitRequestMessage,
-    InitResponseMessage,
-    CallRequestMessage,
-    CallResponseMessage,
-    PingRequestMessage,
-    PingResponseMessage,
-    ErrorMessage,
-]
-
+from .messages import ALL_MESSAGES
 MESSAGE_TYPES_TO_CLASSES = {
     msg.message_type: msg
     for msg in ALL_MESSAGES
