+++ tchannel/messages/common.py
@@ -1,0 +1,1 @@
+PROTOCOL_VERSION = 0x02
